requirements.txt
discord.py
psutil
aiohttp
youtube_dl
discord-components
pillow
phonenumbers
requests